app/requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
gunicorn
whitenoise>=6.0
django-import-export

